<!-- If necessary, assign reviewers that know the area or changes well. Feel free to tag any additional reviewers you see fit. -->

### Details
<!-- Explanation of the change or anything fishy that is going on -->

This PR introduces a new configuration to `Onyx.init()` – the option to specify skippable collection member IDs. It will be possible to pass an array of collection member IDs (strings) which updates will be ignored when using Onyx methods. Additionally, any subscribers from these keys to won't receive any data from Onyx.

Example:

```tsx
// Onyx initialisation
Onyx.init({
    keys: ONYXKEYS,
    skippableCollectionMemberIDs: ['0', '-1', 'undefined', 'null', 'NaN'],
});

// Now, any updates to a collection member that has the specified IDs ('0', '-1', 'undefined', 'null', 'NaN')
// will be ignored and won't produce any effect in Onyx cache/storage.
// Onyx subscribes won't receive updates about these IDs.

// Let's assume it's undefined.
const reportID = undefined;

// Result key is `report_undefined`. This operation won't produce any changes because the collection member ID `undefined`
// was configured to be skipped.
Onyx.merge(`${ONYXKEYS.COLLECTION.REPORT}${reportID}`, {...});

// useOnyx will never receive a report, even if it was stored previously by mistake in a `report_undefined` key.
const [report] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${reportID}`);
```

### Related Issues
<!-- Please replace GH_LINK with the link to the GitHub issue this Pull Request is related to -->
https://github.com/Expensify/App/issues/50360

### Automated Tests

Unit tests were created to cover the changes.

### Manual Tests

General tests in E/App:

1. Login with an account.
2. Go to a chat and send a message. React to that message.
3. Create a thread from that message and send a message inside.
4. Create a task in the chat. Go to its details and mark it as complete.
5. Go to FAB -> Create Expense. Send it to someone. Go to Reports tab and access the expense's details.
6. Go to Settings and logout the user.

### Author Checklist

- [x] I linked the correct issue in the `### Related Issues` section above
- [x] I wrote clear testing steps that cover the changes made in this PR
    - [x] I added steps for local testing in the `Tests` section
    - [x] I tested this PR with a [High Traffic account](https://github.com/Expensify/App/blob/main/contributingGuides/CONTRIBUTING.md#high-traffic-accounts) against the staging or production API to ensure there are no regressions (e.g. long loading states that impact usability).
- [x] I included screenshots or videos for tests on [all platforms](https://github.com/Expensify/App/blob/main/contributingGuides/CONTRIBUTING.md#make-sure-you-can-test-on-all-platforms)
- [x] I ran the tests on **all platforms** & verified they passed on:
    - [x] Android / native
    - [x] Android / Chrome
    - [x] iOS / native
    - [x] iOS / Safari
    - [x] MacOS / Chrome / Safari
    - [x] MacOS / Desktop
- [x] I verified there are no console errors (if there's a console error not related to the PR, report it or open an issue for it to be fixed)
- [x] I followed proper code patterns (see [Reviewing the code](https://github.com/Expensify/App/blob/main/contributingGuides/PR_REVIEW_GUIDELINES.md#reviewing-the-code))
    - [x] I verified that any callback methods that were added or modified are named for what the method does and never what callback they handle (i.e. `toggleReport` and not `onIconClick`)
    - [x] I verified that the left part of a conditional rendering a React component is a boolean and NOT a string, e.g. `myBool && <MyComponent />`.
    - [x] I verified that comments were added to code that is not self explanatory
    - [x] I verified that any new or modified comments were clear, correct English, and explained "why" the code was doing something instead of only explaining "what" the code was doing.
    - [x] I verified proper file naming conventions were followed for any new files or renamed files. All non-platform specific files are named after what they export and are not named "index.js". All platform-specific files are named for the platform the code supports as outlined in the README.
    - [x] I verified the JSDocs style guidelines (in [`STYLE.md`](https://github.com/Expensify/App/blob/main/contributingGuides/STYLE.md#jsdocs)) were followed
- [x] If a new code pattern is added I verified it was agreed to be used by multiple Expensify engineers
- [x] I followed the guidelines as stated in the [Review Guidelines](https://github.com/Expensify/App/blob/main/contributingGuides/PR_REVIEW_GUIDELINES.md)
- [x] I tested other components that can be impacted by my changes (i.e. if the PR modifies a shared library or component like `Avatar`, I verified the components using `Avatar` are working as expected)
- [x] I verified all code is DRY (the PR doesn't include any logic written more than once, with the exception of tests)
- [x] I verified any variables that can be defined as constants (ie. in CONST.js or at the top of the file that uses the constant) are defined as such
- [x] I verified that if a function's arguments changed that all usages have also been updated correctly
- [x] If a new component is created I verified that:
    - [x] A similar component doesn't exist in the codebase
    - [x] All props are defined accurately and each prop has a `/** comment above it */`
    - [x] The file is named correctly
    - [x] The component has a clear name that is non-ambiguous and the purpose of the component can be inferred from the name alone
    - [x] The only data being stored in the state is data necessary for rendering and nothing else
    - [x] If we are not using the full Onyx data that we loaded, I've added the proper selector in order to ensure the component only re-renders when the data it is using changes
    - [x] For Class Components, any internal methods passed to components event handlers are bound to `this` properly so there are no scoping issues (i.e. for `onClick={this.submit}` the method `this.submit` should be bound to `this` in the constructor)
    - [x] Any internal methods bound to `this` are necessary to be bound (i.e. avoid `this.submit = this.submit.bind(this);` if `this.submit` is never passed to a component event handler like `onClick`)
    - [x] All JSX used for rendering exists in the render method
    - [x] The component has the minimum amount of code necessary for its purpose, and it is broken down into smaller components in order to separate concerns and functions
- [x] If any new file was added I verified that:
    - [x] The file has a description of what it does and/or why is needed at the top of the file if the code is not self explanatory
- [x] If the PR modifies a generic component, I tested and verified that those changes do not break usages of that component in the rest of the App (i.e. if a shared library or component like `Avatar` is modified, I verified that `Avatar` is working as expected in all cases)
- [x] If the `main` branch was merged into this PR after a review, I tested again and verified the outcome was still expected according to the `Test` steps.
- [x] I have checked off every checkbox in the PR author checklist, including those that don't apply to this PR.

### Screenshots/Videos
<details>
<summary>Android: Native</summary>


https://github.com/user-attachments/assets/0f5a4cf7-0c32-4ab5-b612-75bb3c3174e6



</details>

<details>
<summary>Android: mWeb Chrome</summary>


https://github.com/user-attachments/assets/317d2606-e986-4483-801e-c8a40704db2c



</details>

<details>
<summary>iOS: Native</summary>


https://github.com/user-attachments/assets/286d3023-cd72-4680-ba28-12db3629c9ae



</details>

<details>
<summary>iOS: mWeb Safari</summary>


https://github.com/user-attachments/assets/8ed500d3-28c9-4a6c-8926-a7ded33fe32b



</details>

<details>
<summary>MacOS: Chrome / Safari</summary>


https://github.com/user-attachments/assets/85565fdf-ba3d-4078-aa40-4510e2fb29d1


https://github.com/user-attachments/assets/e236c798-3f1b-45d8-bc7a-cc6181b45e85



</details>

<details>
<summary>MacOS: Desktop</summary>


https://github.com/user-attachments/assets/8d6b4fec-8cf4-423d-938b-fe24608a650b



</details>