@fabioh8010 2025-01-13
I've just profiled the Onyx changes -- no significant changes in performance ✅ 

---

# Profiling

## main Onyx branch

### Startup / Open Big Report

* Render time (ms): 7373 / 7350 / 7380

### Switch to Big Report

* Render time (ms): 4392 / 4373 / 4470

### Send Message

* Render time (ms): 1340 / 1246 / 1247

### Create Expense Flow

* Render time (ms): 6025 / 6103 / 6550

## skippable member IDs Onyx branch

### Open Big Report

* Render time (ms): 7350 / 7432 / 7368

### Switch to Big Report

* Render time (ms): 4359 / 4216 / 4343

### Send Message

* Render time (ms): 1379 / 1295 / 1401

### Create Expense Flow

* Render time (ms): 5902 / 5840 / 5910

---

# Reassure

 - **Current**: feature/onyx-skippable-collection-member-ids (4b09608cb716f0c64219d00f3c735a69eceb1b31) - 2025-01-13 21:56:57Z
 - **Baseline**: feature/onyx-skippable-collection-member-ids (4b09608cb716f0c64219d00f3c735a69eceb1b31) - 2025-01-13 21:48:01Z

### Significant Changes To Duration

| Name                                                                                    | Type     | Duration                                | Count  |
| --------------------------------------------------------------------------------------- | -------- | --------------------------------------- | ------ |
| OptionsListUtils [OptionsListUtils] getShareLogOptions                                  | function | 37.6 ms → 33.1 ms (-4.5 ms, -11.9%)  🟢 | 1 → 1  |
| OptionsListUtils [OptionsListUtils] getSearchOptions                                    | function | 67.4 ms → 62.4 ms (-5.1 ms, -7.5%)  🟢  | 1 → 1  |
| OptionsListUtils [OptionsListUtils] getShareDestinationOptions                          | function | 61.3 ms → 56.1 ms (-5.1 ms, -8.4%)  🟢  | 1 → 1  |
| SidebarUtils [SidebarUtils] getOrderedReportIDs on 15k reports for default priorityMode | function | 209.3 ms → 201.5 ms (-7.8 ms, -3.7%)    | 1 → 1  |
<details>
<summary>Show details</summary>

| Name                                                                                    | Type     | Duration                                                                                                                                                                                                                                                              | Count                                                                                                                                                         |
| --------------------------------------------------------------------------------------- | -------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------- |
| OptionsListUtils [OptionsListUtils] getShareLogOptions                                  | function | **Baseline**<br/>Mean: 37.6 ms<br/>Stdev: 2.1 ms (5.5%)<br/>Runs: 36.9 41.9 37.5 36.2 39.6 35.9 39.0 36.1 35.2 37.6<br/><br/>**Current**<br/>Mean: 33.1 ms<br/>Stdev: 1.1 ms (3.4%)<br/>Runs: 32.6 32.3 33.6 32.0 32.3 35.5 34.4 32.5 33.5 32.4                       | **Baseline**<br/>Mean: 1<br/>Stdev: 0 (0.0%)<br/>Runs: 1 1 1 1 1 1 1 1 1 1<br/><br/>**Current**<br/>Mean: 1<br/>Stdev: 0 (0.0%)<br/>Runs: 1 1 1 1 1 1 1 1 1 1 |
| OptionsListUtils [OptionsListUtils] getSearchOptions                                    | function | **Baseline**<br/>Mean: 67.4 ms<br/>Stdev: 2.3 ms (3.4%)<br/>Runs: 65.0 69.4 64.7 68.4 65.1 68.8 66.7 69.3 65.6 71.3<br/><br/>**Current**<br/>Mean: 62.4 ms<br/>Stdev: 1.3 ms (2.1%)<br/>Runs: 61.3 63.3 61.1 63.4 63.4 63.9 60.8 64.0 61.5 61.2                       | **Baseline**<br/>Mean: 1<br/>Stdev: 0 (0.0%)<br/>Runs: 1 1 1 1 1 1 1 1 1 1<br/><br/>**Current**<br/>Mean: 1<br/>Stdev: 0 (0.0%)<br/>Runs: 1 1 1 1 1 1 1 1 1 1 |
| OptionsListUtils [OptionsListUtils] getShareDestinationOptions                          | function | **Baseline**<br/>Mean: 61.3 ms<br/>Stdev: 2.1 ms (3.5%)<br/>Runs: 64.0 59.2 59.4 63.1 61.8 63.6 59.2 63.7 59.3 59.7<br/><br/>**Current**<br/>Mean: 56.1 ms<br/>Stdev: 1.9 ms (3.3%)<br/>Runs: 54.7 57.3 54.4 57.3 54.7 55.3 59.9 56.4 57.4 53.9                       | **Baseline**<br/>Mean: 1<br/>Stdev: 0 (0.0%)<br/>Runs: 1 1 1 1 1 1 1 1 1 1<br/><br/>**Current**<br/>Mean: 1<br/>Stdev: 0 (0.0%)<br/>Runs: 1 1 1 1 1 1 1 1 1 1 |
| SidebarUtils [SidebarUtils] getOrderedReportIDs on 15k reports for default priorityMode | function | **Baseline**<br/>Mean: 209.3 ms<br/>Stdev: 2.0 ms (0.9%)<br/>Runs: 210.7 211.5 208.0 206.0 211.2 207.2 211.2 207.7 210.5 209.1<br/><br/>**Current**<br/>Mean: 201.5 ms<br/>Stdev: 1.8 ms (0.9%)<br/>Runs: 201.3 199.9 203.8 200.2 205.1 200.8 201.7 202.4 199.2 200.9 | **Baseline**<br/>Mean: 1<br/>Stdev: 0 (0.0%)<br/>Runs: 1 1 1 1 1 1 1 1 1 1<br/><br/>**Current**<br/>Mean: 1<br/>Stdev: 0 (0.0%)<br/>Runs: 1 1 1 1 1 1 1 1 1 1 |
</details>



### Meaningless Changes To Duration

<details>
<summary>Show entries</summary>

| Name                                                                                                                                                                    | Type     | Duration                              | Count    |
| ----------------------------------------------------------------------------------------------------------------------------------------------------------------------- | -------- | ------------------------------------- | -------- |
| [BaseOptionsList] Should press a list item                                                                                                                              | render   | 0.0 ms → 0.0 ms                       | 4 → 4    |
| [BaseOptionsList] Should render 1 section and a thousand items                                                                                                          | render   | 0.0 ms → 0.0 ms                       | 1 → 1    |
| [BaseOptionsList] Should render multiple selection and select 4 items                                                                                                   | render   | 0.0 ms → 0.0 ms                       | 13 → 13  |
| [BaseOptionsList] Should scroll and select a few items                                                                                                                  | render   | 0.0 ms → 0.0 ms                       | 14 → 14  |
| [GooglePlacesUtils] getAddressComponents on a big dataset                                                                                                               | function | 0.0 ms → 0.0 ms (+0.0 ms, +3.9%)      | 1 → 1    |
| [ModifiedExpenseMessage] getForReportAction on 1k reports and policies                                                                                                  | function | 0.1 ms → 0.1 ms (+0.0 ms, +3.2%)      | 1 → 1    |
| [ReportActionCompose] should press create button                                                                                                                        | render   | 0.0 ms → 0.0 ms                       | 13 → 13  |
| [ReportActionCompose] should press send message button                                                                                                                  | render   | 0.0 ms → 0.0 ms                       | 12 → 12  |
| [ReportActionCompose] should render Composer with text input interactions                                                                                               | render   | 0.0 ms → 0.0 ms                       | 13 → 13  |
| [ReportActionsList] should render ReportActionsList with 500 reportActions stored                                                                                       | render   | 0.0 ms → 0.0 ms                       | 7 → 7    |
| [SearchRouter] should react to text input changes                                                                                                                       | render   | 0.0 ms → 0.0 ms                       | 11 → 11  |
| [SearchRouter] should render list with cached options                                                                                                                   | render   | 0.0 ms → 0.0 ms                       | 9 → 9    |
| [SelectionList] should press a list item                                                                                                                                | render   | 0.0 ms → 0.0 ms                       | 4 → 4    |
| [SelectionList] should render 1 section and a thousand items                                                                                                            | render   | 0.0 ms → 0.0 ms                       | 1 → 1    |
| [SelectionList] should render multiple selection and select 3 items                                                                                                     | render   | 0.0 ms → 0.0 ms                       | 8 → 8    |
| [SelectionList] should scroll and select a few items                                                                                                                    | render   | 0.0 ms → 0.0 ms                       | 12 → 12  |
| OptionsListUtils [OptionsListUtils] empty search term with selected options and mocked personal details                                                                 | function | 3.5 ms → 3.3 ms (-0.3 ms, -7.2%)      | 1 → 1    |
| OptionsListUtils [OptionsListUtils] getFilteredOptions with search value                                                                                                | function | 24.6 ms → 22.0 ms (-2.7 ms, -10.9%)   | 1 → 1    |
| OptionsListUtils [OptionsListUtils] getMemberInviteOptions                                                                                                              | function | 2.0 ms → 2.3 ms (+0.3 ms, +14.6%)     | 1 → 1    |
| OptionsListUtils [OptionsListUtils] worst case scenario with a search term that matches a subset of selectedOptions, filteredRecentReports, and filteredPersonalDetails | function | 15.9 ms → 15.7 ms (-0.2 ms, -1.0%)    | 1 → 1    |
| PolicyUtils getMemberAccountIDsForWorkspace 500 policy members with errors and personal details                                                                         | function | 0.1 ms → 0.1 ms (+0.0 ms, +3.8%)      | 1 → 1    |
| PolicyUtils getMemberAccountIDsForWorkspace 500 policy members with personal details                                                                                    | function | 0.1 ms → 0.1 ms (-0.0 ms, -8.2%)      | 1 → 1    |
| ReportActionsUtils [ReportActionsUtils] getLastClosedReportAction on 10k ReportActions                                                                                  | function | 0.6 ms → 0.6 ms (-0.0 ms, -1.8%)      | 1 → 1    |
| ReportActionsUtils [ReportActionsUtils] getLastVisibleAction on 10k reportActions                                                                                       | function | 12.0 ms → 12.5 ms (+0.5 ms, +3.8%)    | 1 → 1    |
| ReportActionsUtils [ReportActionsUtils] getLastVisibleAction on 10k reportActions with actionsToMerge                                                                   | function | 12.8 ms → 12.7 ms (-0.0 ms, ±0.0%)    | 1 → 1    |
| ReportActionsUtils [ReportActionsUtils] getLastVisibleMessage on 10k ReportActions                                                                                      | function | 12.5 ms → 12.8 ms (+0.3 ms, +2.8%)    | 1 → 1    |
| ReportActionsUtils [ReportActionsUtils] getLastVisibleMessage on 10k ReportActions with actionsToMerge                                                                  | function | 12.8 ms → 13.2 ms (+0.4 ms, +3.1%)    | 1 → 1    |
| ReportActionsUtils [ReportActionsUtils] getMostRecentIOURequestActionID on 10k ReportActions                                                                            | function | 0.0 ms → 0.0 ms (+0.0 ms, +43.5%) 🔴  | 1 → 1    |
| ReportActionsUtils [ReportActionsUtils] getMostRecentReportActionLastModified                                                                                           | function | 2.0 ms → 2.2 ms (+0.2 ms, +9.9%)      | 1 → 1    |
| ReportActionsUtils [ReportActionsUtils] getSortedReportActionsForDisplay on 10k ReportActions                                                                           | function | 0.6 ms → 0.6 ms (+0.0 ms, +6.6%)      | 1 → 1    |
| ReportUtils [ReportUtils] canDeleteReportAction on 1k reports and policies                                                                                              | function | 0.0 ms → 0.0 ms (-0.0 ms, -29.9%) 🟢  | 1 → 1    |
| ReportUtils [ReportUtils] canShowReportRecipientLocalTime on 1k participants                                                                                            | function | 0.0 ms → 0.0 ms (+0.0 ms, +14.3%)     | 1 → 1    |
| ReportUtils [ReportUtils] findLastAccessedReport on 2k reports and policies                                                                                             | function | 0.2 ms → 0.3 ms (+0.0 ms, +13.9%)     | 1 → 1    |
| ReportUtils [ReportUtils] getDisplayNamesWithTooltips 1k participants                                                                                                   | function | 1.1 ms → 1.5 ms (+0.4 ms, +38.0%) 🔴  | 1 → 1    |
| ReportUtils [ReportUtils] getIcons on 1k participants                                                                                                                   | function | 0.0 ms → 0.1 ms (+0.0 ms, +348.9%) 🔴 | 1 → 1    |
| ReportUtils [ReportUtils] getIconsForParticipants on 1k participants                                                                                                    | function | 0.9 ms → 0.9 ms (+0.0 ms, +1.1%)      | 1 → 1    |
| ReportUtils [ReportUtils] getIOUReportActionDisplayMessage on 1k policies                                                                                               | function | 0.0 ms → 0.0 ms (-0.0 ms, -5.7%)      | 1 → 1    |
| ReportUtils [ReportUtils] getMoneyRequestOptions on 1k participants                                                                                                     | function | 0.0 ms → 0.0 ms (-0.0 ms, -8.1%)      | 1 → 1    |
| ReportUtils [ReportUtils] getReportName on 1k participants                                                                                                              | function | 0.0 ms → 0.0 ms (+0.0 ms, +126.6%) 🔴 | 1 → 1    |
| ReportUtils [ReportUtils] getReportPreviewMessage on 1k policies                                                                                                        | function | 0.1 ms → 0.1 ms (-0.0 ms, -2.6%)      | 1 → 1    |
| ReportUtils [ReportUtils] getReportRecipientAccountID on 1k participants                                                                                                | function | 0.0 ms → 0.0 ms (+0.0 ms, +927.9%) 🔴 | 1 → 1    |
| ReportUtils [ReportUtils] getTransactionDetails on 1k reports                                                                                                           | function | 0.1 ms → 0.1 ms (+0.0 ms, +4.4%)      | 1 → 1    |
| ReportUtils [ReportUtils] getWorkspaceChat on 1k policies                                                                                                               | function | 0.1 ms → 0.2 ms (+0.0 ms, +2.4%)      | 1 → 1    |
| ReportUtils [ReportUtils] getWorkspaceIcon on 1k policies                                                                                                               | function | 0.0 ms → 0.0 ms (-0.0 ms, -25.4%) 🟢  | 1 → 1    |
| ReportUtils [ReportUtils] shouldReportBeInOptionList on 1k participant                                                                                                  | function | 0.0 ms → 0.0 ms (+0.0 ms, +364.8%) 🔴 | 1 → 1    |
| SidebarLinks [SidebarLinks] should click on list item                                                                                                                   | render   | 0.0 ms → 0.0 ms                       | 9 → 9    |
| SidebarLinks [SidebarLinks] should render Sidebar with 500 reports stored                                                                                               | render   | 0.0 ms → 0.0 ms                       | 8 → 8    |
| SidebarUtils [SidebarUtils] getOptionData                                                                                                                               | function | 1.6 ms → 1.5 ms (-0.1 ms, -5.1%)      | 1 → 1    |
| SidebarUtils [SidebarUtils] getOrderedReportIDs on 15k reports for GSD priorityMode                                                                                     | function | 195.5 ms → 192.2 ms (-3.3 ms, -1.7%)  | 1 → 1    |
</details>


<details>
<summary>Show details</summary>

| Name                                                                                                                                                                    | Type     | Duration                                                                                                                                                                                                                                                                                                    | Count                                                                                                                                                                                                 |
| ----------------------------------------------------------------------------------------------------------------------------------------------------------------------- | -------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
| [BaseOptionsList] Should press a list item                                                                                                                              | render   | **Baseline**<br/>Mean: 0.0 ms<br/>Stdev: 0.0 ms (NaN%)<br/>Runs: 0 0 0 0 0 0 0 0 0 0<br/><br/>**Current**<br/>Mean: 0.0 ms<br/>Stdev: 0.0 ms (NaN%)<br/>Runs: 0 0 0 0 0 0 0 0 0 0                                                                                                                           | **Baseline**<br/>Mean: 4<br/>Stdev: 0 (0.0%)<br/>Runs: 4 4 4 4 4 4 4 4 4 4<br/><br/>**Current**<br/>Mean: 4<br/>Stdev: 0 (0.0%)<br/>Runs: 4 4 4 4 4 4 4 4 4 4                                         |
…